+++ CHANGELOG.md
@@ -4,6 +4,15 @@ The changelog for [Kommunicate-iOS-SDK](https://github.com/Kommunicate-io/Kommun
 
 ## [Unreleased]
 
+
+## [6.3.0] - 2021-09-30
+
+### Enhancements
+- Update ApplozicSwift to 6.4.0
+- Added support for Xcode 13 and iOS 15
+- [TD-2111] Added a password field to PreChat Form View.
+- [TD-2099] Add support to update team for existing conversation
+
 ## [6.2.1] - 2021-08-19
 
 ### Enhancements
@@ -12,6 +21,7 @@ The changelog for [Kommunicate-iOS-SDK](https://github.com/Kommunicate-io/Kommun
 
 ## [6.2.0] - 2021-08-18Z
 
+## [6.2.0] - 2021-08-18Z
 ### Fixes
 
 - [TD-1943] Fix for launch conversation error
@@ -19,6 +29,7 @@ The changelog for [Kommunicate-iOS-SDK](https://github.com/Kommunicate-io/Kommun
 ### Enhancements
 
 - [TD-1867] Updated bot detail API
+- [TD-1731] Added support to sync messages on chat screen opening
 
 ## [6.1.0] - 2021-06-22